Support updates for each object type
