+++ src/WP_CLI/CommandWithTranslation.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace WP_CLI;
+use WP_CLI\Formatter;
 
 /**
  * Base class for WP-CLI commands that deal with translations
@@ -122,83 +123,11 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 	}
 
 	/**
-	 * Installs a given language.
-	 *
-	 * Downloads the language pack from WordPress.org.
-	 *
-	 * ## OPTIONS
-	 *
-	 * <language>...
-	 * : Language code to install.
-	 *
-	 * [--activate]
-	 * : If set, the language will be activated immediately after install.
-	 *
-	 * ## EXAMPLES
-	 *
-	 *     # Install the Japanese language.
-	 *     $ wp language core install ja
-	 *     Success: Language installed.
-	 *
-	 * @subcommand install
-	 */
-	public function install( $args, $assoc_args ) {
-
-		$language_codes = $args;
-
-		if(  1 < count( $language_codes )  &&  in_array( true , $assoc_args , true ) ){
-			\WP_CLI::error( 'Only a single language can be active.' );
-		}
-
-		$available = $this->get_installed_languages();
-
-		foreach ($language_codes as $language_code) {
-
-			if ( in_array( $language_code, $available, true ) ) {
-				\WP_CLI::warning( "Language '{$language_code}' already installed." );
-			} else {
-				$response = $this->download_language_pack( $language_code );
-
-				if ( is_wp_error( $response ) ) {
-					\WP_CLI::error( $response );
-				} else {
-					\WP_CLI::success( "Language installed." );
-				}
-			}
-
-			if ( \WP_CLI\Utils\get_flag_value( $assoc_args, 'activate' ) ) {
-				$this->activate( array( $language_code ), array() );
-			}
-		}
-
-	}
-
-	/**
-	 * Updates installed languages.
-	 *
-	 * Updates installed languages for core, plugins and themes.
-	 *
-	 * ## OPTIONS
-	 *
-	 * [--dry-run]
-	 * : Preview which translations would be updated.
-	 *
-	 * ## EXAMPLES
-	 *
-	 *     $ wp language core update
-	 *     Updating 'Japanese' translation for Akismet 3.1.11...
-	 *     Downloading translation from https://downloads.wordpress.org/translation/plugin/akismet/3.1.11/ja.zip...
-	 *     Translation updated successfully.
-	 *     Updating 'Japanese' translation for Twenty Fifteen 1.5...
-	 *     Downloading translation from https://downloads.wordpress.org/translation/theme/twentyfifteen/1.5/ja.zip...
-	 *     Translation updated successfully.
-	 *     Success: Updated 2/2 translations.
-	 *
-	 * @subcommand update
+	 * Updates installed languages for the current object type.
 	 */
 	public function update( $args, $assoc_args ) {
-
 		$updates = $this->get_translation_updates();
+
 		if ( empty( $updates ) ) {
 			\WP_CLI::success( 'Translations are up to date.' );
 
@@ -208,13 +137,15 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 		// Gets a list of all languages.
 		$all_languages = $this->get_all_languages();
 
+		$updates_per_type = array();
+
 		// Formats the updates list.
 		foreach ( $updates as $update ) {
-			if ( 'plugin' == $update->type ) {
+			if ( 'plugin' === $update->type ) {
 				$plugins	 = get_plugins( '/' . $update->slug );
 				$plugin_data = array_shift( $plugins );
 				$name		 = $plugin_data['Name'];
-			} elseif ( 'theme' == $update->type ) {
+			} elseif ( 'theme' === $update->type ) {
 				$theme_data	 = wp_get_theme( $update->slug );
 				$name		 = $theme_data['Name'];
 			} else { // Core
@@ -231,12 +162,19 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 			$update->Name		 = $name;
 			$update->Version	 = $update->version;
 			$update->Language	 = $translation->english_name;
+
+			$updates_per_type[ $update->type ] = $update;
 		}
+
+		$obj_type = rtrim( $this->obj_type, 's' );
+		$available_updates = $updates_per_type[ $obj_type ];
+
+		$num_to_update	 = count( $available_updates );
 
 		// Only preview which translations would be updated.
 		if ( \WP_CLI\Utils\get_flag_value( $assoc_args, 'dry-run' ) ) {
-			\WP_CLI::line( sprintf( 'Available %d translations updates:', count( $updates ) ) );
-			\WP_CLI\Utils\format_items( 'table', $updates, array( 'Type', 'Name', 'Version', 'Language' ) );
+			\WP_CLI::line( sprintf( 'Available %d translations updates:', $num_to_update ) );
+			\WP_CLI\Utils\format_items( 'table', $available_updates, array( 'Type', 'Name', 'Version', 'Language' ) );
 
 			return;
 		}
@@ -245,7 +183,7 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 		$results = array();
 
 		// Update translations.
-		foreach ( $updates as $update ) {
+		foreach ( $available_updates as $update ) {
 			\WP_CLI::line( "Updating '{$update->Language}' translation for {$update->Name} {$update->Version}..." );
 
 			$result = Utils\get_upgrader( $upgrader )->upgrade( $update );
@@ -253,12 +191,11 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 			$results[] = $result;
 		}
 
-		$num_to_update	 = count( $updates );
-		$num_updated	 = count( array_filter( $results ) );
+		$num_updated = count( array_filter( $results ) );
 
 		$line = "Updated $num_updated/$num_to_update translations.";
 
-		if ( $num_to_update == $num_updated ) {
+		if ( $num_to_update === $num_updated ) {
 			\WP_CLI::success( $line );
 		} else if ( $num_updated > 0 ) {
 			\WP_CLI::warning( $line );
@@ -275,22 +212,39 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 	 */
 	protected function get_translation_updates() {
 		$available = $this->get_installed_languages();
+
 		$func = function() use ( $available ) {
 			return $available;
 		};
-		$filters = array( 'plugins_update_check_locales', 'themes_update_check_locales' );
-		foreach( $filters as $filter ) {
-			add_filter( $filter, $func );
+
+		switch( $this->obj_type ) {
+			case 'plugins':
+				add_filter( 'plugins_update_check_locales', $func );
+
+				wp_clean_plugins_cache();
+				// Check for Plugin translation updates.
+				wp_update_plugins();
+
+				remove_filter( 'plugins_update_check_locales', $func );
+				break;
+			case 'themes':
+				add_filter( 'themes_update_check_locales', $func );
+
+				wp_clean_themes_cache();
+				// Check for Theme translation updates.
+				wp_update_plugins();
+
+				remove_filter( 'themes_update_check_locales', $func );
+				break;
+			default:
+				delete_site_transient( 'update_core' );
+
+				// Check for Core translation updates.
+				wp_version_check();
 		}
-		$this->wp_clean_update_cache(); // Clear existing update caches.
-		wp_version_check();      // Check for Core translation updates.
-		wp_update_themes();      // Check for Theme translation updates.
-		wp_update_plugins();     // Check for Plugin translation updates.
-		foreach( $filters as $filter ) {
-			remove_filter( $filter, $func );
-		}
-		$updates = wp_get_translation_updates(); // Retrieves a list of all translations updates available.
-		return $updates;
+
+		// Returns a list of all translations updates available.
+		return wp_get_translation_updates();
 	}
 
 	/**
@@ -300,7 +254,7 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 	 *
 	 * @param string $download Language code to download.
 	 * @param string $slug Plugin or theme slug. Not used for core.
-	 * @return string|WP_Error Returns the language code if successfully downloaded, or a WP_Error object on failure.
+	 * @return string|\WP_Error Returns the language code if successfully downloaded, or a WP_Error object on failure.
 	 */
 	protected function download_language_pack( $download, $slug = null ) {
 
@@ -348,7 +302,7 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 	/**
 	 * Return a list of all languages
 	 *
-	 * @param string $slug Plugin or theme slug. Not used fore core.
+	 * @param string $slug Optional. Plugin or theme slug. Not used fore core.
 	 *
 	 * @return array
 	 */
@@ -379,91 +333,12 @@ abstract class CommandWithTranslation extends \WP_CLI_Command {
 	}
 
 	/**
-	 * Uninstalls a given language.
-	 *
-	 * ## OPTIONS
-	 *
-	 * <language>...
-	 * : Language code to uninstall.
-	 *
-	 * ## EXAMPLES
-	 *
-	 *     $ wp language core uninstall ja
-	 *     Success: Language uninstalled.
-	 *
-	 * @subcommand uninstall
-	 */
-	public function uninstall( $args, $assoc_args ) {
-		global $wp_filesystem;
-
-		$language_codes = $args;
-
-		$available = $this->get_installed_languages();
-
-		foreach ($language_codes as $language_code) {
-
-			if ( ! in_array( $language_code, $available ) ) {
-				\WP_CLI::error( "Language not installed." );
-			}
-
-			$dir = 'core' === $this->obj_type ? '' : "/$this->obj_type";
-			$files = scandir( WP_LANG_DIR . $dir );
-			if ( ! $files ) {
-				\WP_CLI::error( "No files found in language directory." );
-			}
-
-			$current_locale = get_locale();
-			if ( $language_code === $current_locale ) {
-				\WP_CLI::warning( "The '{$language_code}' language is active." );
-				exit;
-			}
-
-			// As of WP 4.0, no API for deleting a language pack
-			WP_Filesystem();
-			$deleted = false;
-			foreach ( $files as $file ) {
-				if ( '.' === $file[0] || is_dir( $file ) ) {
-					continue;
-				}
-				$extension_length = strlen( $language_code ) + 4;
-				$ending = substr( $file, -$extension_length );
-				if ( ! in_array( $file, array( $language_code . '.po', $language_code . '.mo' ) ) && ! in_array( $ending, array( '-' . $language_code . '.po', '-' . $language_code . '.mo' ) ) ) {
-					continue;
-				}
-				$deleted = $wp_filesystem->delete( WP_LANG_DIR . $dir . '/' . $file );
-			}
-
-			if ( $deleted ) {
-				\WP_CLI::success( "Language uninstalled." );
-			} else {
-				\WP_CLI::error( "Couldn't uninstall language." );
-			}
-
-		}
-
-	}
-
-	/**
 	 * Get Formatter object based on supplied parameters.
 	 *
 	 * @param array $assoc_args Parameters passed to command. Determines formatting.
-	 * @return \WP_CLI\Formatter
+	 * @return Formatter
 	 */
 	protected function get_formatter( &$assoc_args ) {
-		return new \WP_CLI\Formatter( $assoc_args, $this->obj_fields, $this->obj_type );
+		return new Formatter( $assoc_args, $this->obj_fields, $this->obj_type );
 	}
-
-	/**
-	 * Replicates wp_clean_update_cache() for use in WP 4.0
-	 */
-	private static function wp_clean_update_cache() {
-		if ( function_exists( 'wp_clean_plugins_cache' ) ) {
-			wp_clean_plugins_cache();
-		} else {
-			delete_site_transient( 'update_plugins' );
-		}
-		wp_clean_themes_cache();
-		delete_site_transient( 'update_core' );
-	}
-
 }
